+++ app/Plugin/Admin/Admin.php
@@ -853,8 +853,8 @@ class Admin {
 		</p>
 		<p>
 			<?php
-			/* translators: %1$s and %2$s will be replaced by external URLs. */
-			echo wp_kses_post( sprintf( __( 'Check out our repository on <a href="%1$s" target="_blank">GitHub</a>. There you will also find <a href="%2$s" target="_blank">some documentation%2$s</a>.', 'personio-integration-light' ), esc_url( 'https://github.com/threadi/wp-personio-integration-light' ), esc_url( Helper::get_github_documentation_link() ), Helper::get_a11n_window_hint() ) );
+			/* translators: %1$s and %2$s will be replaced by external URLs, %3$s will be replaced by the window-hint. */
+			echo wp_kses_post( sprintf( __( 'Check out our repository on <a href="%1$s" target="_blank">GitHub</a>. There you will also find <a href="%2$s" target="_blank">some documentation%3$s</a>.', 'personio-integration-light' ), esc_url( 'https://github.com/threadi/wp-personio-integration-light' ), esc_url( Helper::get_github_documentation_link() ), Helper::get_a11n_window_hint() ) );
 			?>
 			</p>
 		<?php